refactor: move subcomponent interface to new file
